Added score system
Gonna add some sounds here and there, other than that I think we're good
besides whatever you can get working for the ProGen Mode.  After I
commit the sounds I'm gonna put all the sprite sheets we used in a
folder and upload that to Google Drive and share a link on Docs page,
please do the same with our sounds if needed.  And post that link to
where we got them.